Add prediction function
(Pepper) Add prediction function
(Pepper) Fix all path bug
(Pepper) Add log system
(Pepper) add log_initialize function
(Pepper) Add set_max_pdb_size function
(Pepper) Add set_min_pdb_size function
(Pepper.__init__) Add some log related attributes
(Pepper.__init__) Add remove_pdb_after_scoring parameter
(Pepper.get_pdb_size) Optimize the download progress
(Pepper.mutate_and_scoring) Fix download PDB files bug
(Pepper.mutate_and_scoring) Add remove pdb scripts

+++ pepper/pepper.py
@@ -1,11 +1,17 @@
 from pathlib import Path
 import pandas as pd
 import numpy as np
+import lightgbm as lgb
+import pickle
 import json
 import re
 import os
+import time
+import logging
+import datetime
 from urllib.request import urlretrieve
 from urllib.error import HTTPError
+from urllib.error import URLError
 from pyrosetta import init
 from pyrosetta.io import pose_from_pdb
 from pyrosetta.toolbox.cleaning import cleanATOM
@@ -13,9 +19,11 @@ from pyrosetta.toolbox.mutants import mutate_residue
 from pyrosetta.rosetta.core.scoring import ScoreFunction
 from pyrosetta.rosetta.core.scoring import ScoreType
 from pyrosetta.rosetta.core.scoring import calc_total_sasa
+from sklearn.preprocessing import StandardScaler
+from sklearn.feature_selection import VarianceThreshold
 
 
-class Psept:
+class Pepper:
     __repack_radius = 8
     __probe_radius = 2.2
 
@@ -39,17 +47,26 @@ class Psept:
 
     __whole_column = __wt_columns + __mu_columns + __diff_columns
 
-    def __init__(self, pdb_info_path='pdb_info.json'):
+    def __init__(self, remove_pdb_after_scoring=True, pdb_info_path='pdb_info.json'):
         """
-        Initilize and setting for Psept
+        Initilize and setting for Pepper
         """
+        
+        dt = datetime.datetime.today().strftime("%Y-%m-%d %H:%M:%S")
+        self.__start_time = dt
+        self.log_initialize()
+        self.__main_log.info('Program start')
         self.__data = None
+        self.__module_root_path = Path(__file__).parent
+        self.__remove_pdb_after_scoring = remove_pdb_after_scoring
 
         try:
-            with open(Path(pdb_info_path), 'r') as f:
+            self.__main_log.info('Loading pdb_info.json...')
+            with open(self.__module_root_path / pdb_info_path, 'r') as f:
                 self.__pdb_info = json.load(f)
-            self
+            self.__main_log.info('Load success')
         except FileNotFoundError:
+            self.__main_log.warning('pdb_info.json not found. Create a empty json file.')
             self.__pdb_info = {
                 'size': {'source': '', 'date': '', 'data': ''},
                 'download_failed': {'source': '', 'date': '', 'data': ''},
@@ -60,17 +77,37 @@ class Psept:
         self.__full_pdb_folder = self.__root_pdb_folder / 'full_pdb'
         self.__clean_pdb_folder = self.__root_pdb_folder / 'clean_pdb'
         self.__skip_pdb_list = self.__pdb_info['download_failed']['data'] + self.__pdb_info['load_failed']['data'] + self.__pdb_info['multimodel']['data']
+        self.__max_pdb_size = None
+        self.__min_pdb_size = None
+
+    def log_initialize(self):
+        """
+        Log initialize and format setting
+        """
+        logging.getLogger('').handlers = []
+        logging.basicConfig(
+            format='%(asctime)s %(name)s %(levelname)s %(message)s ',
+            level=logging.INFO, datefmt='%Y-%m-%d %H:%M:%S'
+        )
+        formatter = logging.Formatter('%(asctime)s %(name)s %(levelname)s %(message)s')
+        main_log = logging.getLogger('main')
+        fh = logging.FileHandler(f'{self.__start_time}.log')
+        fh.setFormatter(formatter)
+        main_log.addHandler(fh)
+        self.__main_log = main_log
 
     def load_savs(self, filepath):
         """
-        Load input sav file.
+        Load input sav file
         Input format: {UniProt ID} {mutated position} {wildtype amino acid} {mutated amino acid}
         e.g., P04637 35 L F
         """
         try:
+            self.__main_log.info('Loading input file...')
             with open(Path(filepath)) as f:
                 data = f.read().splitlines()
         except:
+            self.__main_log.warning('Load failed. Program exit.')
             raise TypeError('File loaded error. Please check file path or file format. Only txt or csv file are allowed.')
         self.__data = data
 
@@ -159,6 +196,18 @@ class Psept:
         """
         cls.__probe_radius = new_probe_radius
 
+    def set_max_pdb_size(self, max_size):
+        """
+        Set max PDB size. All PDB which size is larger than this threshold will be excluded.
+        """
+        self.__max_pdb_size = max_size
+
+    def set_min_pdb_size(self, min_size):
+        """
+        Set min PDB size. All PDB which size is smaller than this threshold will be exclueded.
+        """
+        self.__min_pdb_size = min_size
+
     def get_pdb_size(self, pdb):
         """
         Get pdb size. This function follows several steps:
@@ -173,17 +222,32 @@ class Psept:
             5. Return size of PDB
         All failed will be recorded in pdb_info.json and remove corresponding PDB file (either full or clean)
         """
+        self.__main_log.info(f'Start getting the size of {pdb}')
         full_pdb_path = self.__root_pdb_folder / 'full_pdb' / (pdb + '.pdb')
         clean_pdb_path = self.__root_pdb_folder / 'clean_pdb' / (pdb + '.clean.pdb')
         if not os.path.isfile(clean_pdb_path):
-            try:
-                self.download_pdb(pdb)
-            except HTTPError:
+            download_success = False
+            for i in range(10):
+                try:
+                    self.download_pdb(pdb)
+                    download_success = True
+                    break
+                except HTTPError:
+                    self.__main_log.warning(f'Download {pdb} failed')
+                    self.__pdb_info['download_failed']['data'].append(pdb)
+                    os.remove(full_pdb_path)
+                    return None
+                except URLError:
+                    time.sleep(5)
+                    continue
+            if not download_success:
+                self.__main_log.warning(f'Downloading {pdb} failed')
                 self.__pdb_info['download_failed']['data'].append(pdb)
                 os.remove(full_pdb_path)
                 return None
             is_multimodle = self.check_pdb_is_multimodel(full_pdb_path)
             if is_multimodle:
+                self.__main_log.warning(f'{pdb} is multimodel')
                 self.__pdb_info['multimodel']['data'].append(pdb)
                 os.remove(full_pdb_path)
                 return None
@@ -191,6 +255,7 @@ class Psept:
         try:
             pose = pose_from_pdb(str(clean_pdb_path))
         except RuntimeError:
+            self.__main_log.warning(f'Load {pdb} failed')
             self.__pdb_info['load_failed']['data'].append(pdb)
             os.remove(full_pdb_path)
             os.remove(clean_pdb_path)
@@ -203,17 +268,29 @@ class Psept:
         """
         Create ordered pdb list
         """
+        self.__main_log.info('Start creating the ordered PDB list')
+        self.__main_log.info(f'Total number of SAVs from input file: {len(self.__data)}')
+        count = 1
         for sav in self.__data:
+            self.__main_log.debug(f'Current sav: {sav}')
             uniprot, pos = sav.split()[:2]
             pos = int(pos)
             pdb_size_list = []
 
             # Get pdb list for sav gene and process each pdb.
-            pdb_list = self.__pdb_info['uniprot_pdb']['data'][uniprot]
+            try:
+                pdb_list = self.__pdb_info['uniprot_pdb']['data'][uniprot]
+            except KeyError:
+                self.__main_log.warning(f'Create failed ({sav}): No related PDB are found.')
+                self.__sav_to_ordered_pdb.append([sav, 'No related PDB are found.'])
+                continue
+
+            self.__main_log.debug(f'pdb_list: {pdb_list}')
             for pdb in pdb_list:
 
-                # Continue if no pdb were found.
+                # Continue if pdb in skip list.
                 if pdb in self.__skip_pdb_list:
+                    self.__main_log.debug(f'{pdb} is in skip_list')
                     continue
 
                 # Find corresponding dataframe.
@@ -221,6 +298,7 @@ class Psept:
                 # Filter data which wasn't containing mutated position.
                 temp_df = temp_df.loc[(self.__map_df['SP_BEG'] < pos) & (self.__map_df['SP_END'] > pos)]
                 if temp_df.empty:
+                    self.__main_log.debug(f'PDB skip: Mutated position is not in {pdb}')
                     continue
 
                 pdb_series = temp_df.iloc[0]
@@ -232,6 +310,7 @@ class Psept:
 
                 # (pdb_end - pdb_beg) != (uni_end - uni_beg) indicate that wrong positions were recorded.
                 if (pdb_end - pdb_beg) != (uni_end - uni_beg):
+                    self.__main_log.debug(f'PDB skip: Wrong position were recorded in {pdb}')
                     continue
 
                 pos_in_pdb = pos + (pdb_beg - uni_beg)
@@ -239,7 +318,6 @@ class Psept:
                 # Get PDB size from pdb_info
                 try:
                     pdb_size = int(self.__pdb_info['size']['data'][pdb])
-
                 # If not recorded, download PDB.
                 except KeyError:
                     pdb_size = self.get_pdb_size(pdb)
@@ -252,7 +330,8 @@ class Psept:
 
             # Continue if no PDB matched.
             if not pdb_size_list:
-                self.__sav_to_ordered_pdb.append([sav, 'No PDB matched.'])
+                self.__main_log.warning(f'Create faile ({sav}): No PDB are usable after create candidate pdb list.')
+                self.__sav_to_ordered_pdb.append([sav, 'No PDB are usable after create candidate pdb list.'])
                 continue
 
             # Sort by size and order by median.
@@ -261,35 +340,35 @@ class Psept:
             while sort_by_size:
                 median = len(sort_by_size)//2
                 middle_start_list.append(sort_by_size.pop(median))
+            self.__main_log.info(f'Create success {sav} ({count} / {len(self.__data)})')
             self.__sav_to_ordered_pdb.append([sav, middle_start_list])
+            count += 1
 
-    def mutate(self):
+    def mutate_and_scoring(self):
         """
         Mutate each SAVs of input file
         """
-
+        self.__main_log.info('Start mutate and scoring')
         # Initialize and create result dataframe
         result_df_column = [
-            'SAV', 'ordered_PDB_list', 'ordered_list_length', 'predict_status',
+            'SAV', 'predict_status', 'predict_label', 'ordered_PDB_list', 'ordered_list_length',
             'current_PDB', 'current_PDB_info', 'chain', 'position', 'PDB_size', 'chain_length',
         ]
         result_df_column = result_df_column + self.__whole_column + ['wt_SASA', 'mu_SASA', 'diff_SASA']
         result_df = pd.DataFrame(columns=result_df_column)
         result_df['SAV'] = [i[0] for i in self.__sav_to_ordered_pdb]
         result_df['ordered_PDB_list'] = [i[1] for i in self.__sav_to_ordered_pdb]
-
+        loop_round = 0
         while True:
-            # Record "failed" if no more PDB in candidate list
-            result_df.loc[
-                (result_df['ordered_list_length'] == 0) & (result_df['predict_status'] == 'unfinished'),
-                'predict_status'
-            ] = 'failed'
+            loop_round += 1
+            # Record "failed" if no more PDB in candidate list of a SAV
+            result_df.loc[(result_df['ordered_list_length'] == 0) & (result_df['predict_status'] == 'unfinished'), 'predict_status'] = 'failed'
 
             # Extract sub_df of unfinished SAVs
             sub_df = result_df[~result_df.predict_status.isin(['success', 'failed'])]
 
             if sub_df.empty:
-                self.__scoring_result = result_df
+                self.__result = result_df
                 break
 
             # Pop new PDB from ordered_PDB_list and update the information
@@ -300,17 +379,42 @@ class Psept:
             sub_df['position'] = sub_df['current_PDB_info'].map(lambda x: x[3])
             sub_df['ordered_PDB_list'] = sub_df['ordered_PDB_list'].apply(lambda x: x[1:])
             sub_df['ordered_list_length'] = sub_df['ordered_PDB_list'].apply(lambda x: len(x))
-
+            count = 0
             for pdb in sub_df['current_PDB'].unique():
+                count += 1
+                self.__main_log.info(f'Current progress: {count} / {len(sub_df["current_PDB"].unique())}, Round: {loop_round}')
+                # Check if PDB file is exist
+                pdb_usable = False
 
-                # Calculate wildtype PDB score
-                wt_pose = pose_from_pdb(str(self.__clean_pdb_folder / (pdb + '.clean.pdb')))
-                wt_score = np.array(self.pyrosetta_scoring(wt_pose))
-                wt_sum = np.sum(wt_score)
-                wt_score = np.append(wt_score, wt_sum)
+                # Download PDB if not exist
+                if not os.path.isfile(self.__clean_pdb_folder / (pdb + '.clean.pdb')):
+                    for i in range(10):
+                        try:
+                            self.download_pdb(pdb)
+                            break
+                        except HTTPError:
+                            break
+                        except URLError:
+                            time.sleep(5)
+                            continue
+                    cleanATOM(str(self.__full_pdb_folder / (pdb + '.pdb')), str(self.__clean_pdb_folder / (pdb + '.clean.pdb')))
+                try:
+                    # Calculate wildtype PDB score
+                    wt_pose = pose_from_pdb(str(self.__clean_pdb_folder / (pdb + '.clean.pdb')))
+                    wt_score = np.array(self.pyrosetta_scoring(wt_pose))
+                    wt_sum = np.sum(wt_score)
+                    wt_score = np.append(wt_score, wt_sum)
+                    pdb_usable = True
+                except RuntimeError:
+                    pass
 
                 temp_df = sub_df[sub_df['current_PDB'] == pdb]
                 for idx, row in temp_df.iterrows():
+                    # If PDB is not usable, skip
+                    if not pdb_usable:
+                        self.__main_log.debug(f'({temp_df.loc[idx]["SAV"]}) PDB {pdb} not usable, label as unfinished.')
+                        sub_df.at[idx, 'predict_status'] = 'unfinished'
+                        continue
                     chain = row['chain']
                     position_in_pdb = row['position']
                     wt_aa, mu_aa = row['SAV'].split()[2:4]
@@ -320,7 +424,8 @@ class Psept:
 
                     # If position not found or residue in pose not equal to SAV, continue
                     if position_in_pose == 0 or wt_pose.residue(position_in_pose).name1() != wt_aa:
-                        sub_df.set_value(idx, 'predict_status', 'unfinished')
+                        self.__main_log.debug(f'({temp_df.loc[idx]["SAV"]}) position not found or residue in pose not equal to SAV, label as unfinished.')
+                        sub_df.at[idx, 'predict_status'] = 'unfinished'
                         continue
 
                     # Get chain number and chain length
@@ -342,19 +447,23 @@ class Psept:
                     result = wt_score.tolist() + mu_score.tolist() + diff_score.tolist()
 
                     # Successful score, update the predict_status
-                    sub_df.set_value(idx, 'predict_status', 'success')
-                    sub_df.set_value(idx, 'chain_length', chain_length)
-
+                    sub_df.at[idx, 'predict_status'] = 'success'
+                    sub_df.at[idx, 'chain_length'] = chain_length
                     # Calculate score of SASA
                     wt_sasa = calc_total_sasa(wt_pose, self.__probe_radius)
                     mu_sasa = calc_total_sasa(mutant_pose, self.__probe_radius)
-                    sub_df.set_value(idx, "wt_SASA", wt_sasa)
-                    sub_df.set_value(idx, "mu_SASA", mu_sasa)
-                    sub_df.set_value(idx, "diff_SASA", mu_sasa - wt_sasa)
+                    sub_df.at[idx, 'wt_SASA'] = wt_sasa
+                    sub_df.at[idx, 'mu_SASA'] = mu_sasa
+                    sub_df.at[idx, 'diff_SASA'] = mu_sasa - wt_sasa
 
                     for i in range(len(self.__whole_column)):
-                        sub_df.set_value(idx, self.__whole_column[i], result[i])
+                        sub_df.at[idx, self.__whole_column[i]] = result[i]
+                    self.__main_log.info(f'{temp_df.loc[idx]["SAV"]} successful scoring')
                 result_df.update(sub_df)
+                result_df.to_csv('structure_energies_score.csv', index=False)
+                if self.__remove_pdb_after_scoring:
+                    os.remove(self.__root_pdb_folder / 'full_pdb' / (pdb + '.pdb'))
+                    os.remove(self.__root_pdb_folder / 'clean_pdb' / (pdb + '.clean.pdb'))
 
     def pyrosetta_scoring(self, pose):
         """
@@ -412,18 +521,55 @@ class Psept:
         init()
         self.__sav_to_ordered_pdb = []
         # Load position data.
-        self.__map_df = pd.read_csv('uniprot_segments_observed.csv', skiprows=1)
+        self.__main_log.info('Loading uniprot_segments_observed.csv...')
+        try:
+            self.__map_df = pd.read_csv(self.__module_root_path / 'uniprot_segments_observed.csv', skiprows=1)
+        except FileNotFoundError:
+            self.__main_log.warning('Load failed. File not found.')
+            raise FileNotFoundError('File not found. Please put the file in this folder and try again.')
 
         # Generate ordered pdb list for input savs.
         self.create_ordered_pdb_list()
-        self.mutate()
 
+        # Mutate and score SAVs
+        self.mutate_and_scoring()
 
-def main():
-    psept = Psept()
-    psept.load_savs('test_sav.txt')
-    psept.scoring()
+    def prediction(self):
+        """
+        Predict input SAVs using pre-trained lgbm model
+        """
+        features = self.__whole_column + ['wt_SASA', 'mu_SASA', 'diff_SASA']
+        # Load pre-trained model
+        with open(self.__module_root_path / 'lgbm_prediction_model.pkl', 'rb') as f:
+            clf = pickle.load(f)
 
+        with open(self.__module_root_path / 'scaler.pkl', 'rb') as f:
+            scaler = pickle.load(f)
 
-if __name__ == '__main__':
-    main()
+        with open(self.__module_root_path / 'selector.pkl', 'rb') as f:
+            selector = pickle.load(f)
+
+        # Feature engineering
+        # High correlation features will be divided by PDB size
+        divided_by_pdb_size_columns = [
+            "wt_fa_atr", "wt_fa_intra_rep", "wt_fa_sol", "wt_fa_elec", "wt_fa_dun", "wt_ref",
+            "mu_fa_atr", "mu_fa_intra_rep", "mu_fa_sol", "mu_fa_elec", "mu_fa_dun", "mu_ref",
+            "wt_SASA", "mu_SASA"
+        ]
+        for columns in divided_by_pdb_size_columns:
+            self.__result[columns] = self.__result[columns].divide(self.__result['PDB_size'])
+
+        # Scale and selection
+        temp_df = self.__result.drop(columns=['predict_label'])
+        temp_df = temp_df.dropna()
+        not_nan_idx = temp_df.index
+        X_test = temp_df[features]
+        X_test = scaler.transform(X_test)
+        X_test = selector.transform(X_test)
+
+        # Prediction
+        y_pred = clf.predict(X_test)
+        for i in range(len(y_pred)):
+            self.__result.loc[not_nan_idx[i]]['predict_label'] = y_pred[i]
+
+        self.__result.to_csv('predict_result.csv', index=False)
